refactor: consolidate WorkOS URL resolution into utils/urls.ts
Address code review on the env-var cleanup:

- Move getLlmGatewayUrl/getTelemetryUrl from settings.ts into urls.ts so the
  derived endpoints sit next to their base (getWorkOSApiUrl). This removes the
  backwards settings->urls import and makes urls.ts a dependency-free leaf that
  is the single home for WorkOS endpoint resolution. getAuthkitDomain stays in
  settings.ts since it's config-backed, not API-host-derived.
- Normalize the trailing slash once inside getWorkOSApiUrl instead of repeating
  the strip at each derivation site.
- Scope the env-var-catalog guard test's claim to dot-access reads and document
  the forms it doesn't cover (bracket access, destructuring).
- Fix a stale docstring on getCliAuthClientId (it is not env-overridable).
