+++ src/lib/agent-interface.ts
@@ -11,7 +11,8 @@ import { analytics } from '../utils/analytics.js';
 import { INSTALLER_INTERACTION_EVENT_NAME } from './constants.js';
 import { LINTING_TOOLS } from './safe-tools.js';
 import { formatWorkOSCommand } from '../utils/command-invocation.js';
-import { getConfig, getLlmGatewayUrl } from './settings.js';
+import { getConfig } from './settings.js';
+import { getLlmGatewayUrl } from '../utils/urls.js';
 import { getCredentials, hasCredentials } from './credentials.js';
 import { ensureValidToken } from './token-refresh.js';
 import type { InstallerEventEmitter } from './events.js';
