+++ src/lib/settings.ts
@@ -1,5 +1,4 @@
 import { config, version } from '../cli.config.js';
-import { getWorkOSApiUrl } from '../utils/urls.js';
 
 /**
  * Get version from package.json
@@ -56,8 +55,7 @@ export function getConfig(): InstallerConfig {
 }
 
 /**
- * Get the CLI auth client ID.
- * Env var overrides config default.
+ * Get the CLI auth client ID (from config; not env-overridable).
  */
 export function getCliAuthClientId(): string {
   return config.workos.clientId;
@@ -66,23 +64,11 @@ export function getCliAuthClientId(): string {
 /**
  * Get the AuthKit domain.
  * Env var overrides config default.
+ *
+ * Note: WorkOS service endpoints (API, dashboard, LLM gateway, telemetry)
+ * live in utils/urls.ts. AuthKit's domain stays here because it's config-
+ * backed rather than derived from the API host.
  */
 export function getAuthkitDomain(): string {
   return process.env.WORKOS_AUTHKIT_DOMAIN || config.workos.authkitDomain;
-}
-
-/**
- * Get the LLM gateway URL.
- * Derived from the WorkOS API host (override via WORKOS_API_URL).
- */
-export function getLlmGatewayUrl(): string {
-  return `${getWorkOSApiUrl().replace(/\/$/, '')}/llm-gateway`;
-}
-
-/**
- * Get the CLI telemetry URL.
- * Derived from the WorkOS API host (override via WORKOS_API_URL).
- */
-export function getTelemetryUrl(): string {
-  return `${getWorkOSApiUrl().replace(/\/$/, '')}/cli`;
 }
